4._HPSHOP/HPSHOP/requirements.txt
selenium
pytest
allure-pytest
webdriver-manager
openpyxl